Move helper modules into lib directory
There's so many now that the directory separation makes more sense than
the underscore prefix.
