+++ src/lib/moduleResolver.ts
