+++ tsconfig.json
@@ -17,15 +17,15 @@
 	],
 	"files": [
 		"./typings/index.d.ts",
-		"./src/_factories.ts",
-		"./src/_InstanceRegistry.ts",
-		"./src/_makeIdGenerator.ts",
-		"./src/_moduleResolver.ts",
-		"./src/_realizeCustomElements.ts",
-		"./src/_RegistryProvider.ts",
-		"./src/_resolveListenersMap.ts",
 		"./src/createApp.ts",
 		"./src/IdentityRegistry.ts",
+		"./src/lib/factories.ts",
+		"./src/lib/InstanceRegistry.ts",
+		"./src/lib/makeIdGenerator.ts",
+		"./src/lib/moduleResolver.ts",
+		"./src/lib/realizeCustomElements.ts",
+		"./src/lib/RegistryProvider.ts",
+		"./src/lib/resolveListenersMap.ts",
 		"./src/main.ts",
 		"./tests/fixtures/action-factory.ts",
 		"./tests/fixtures/action-instance.ts",
